+++ kodilla-testing/src/main/java/com/kodilla/testing/shape/Shape.java
@@ -2,7 +2,7 @@ package com.kodilla.testing.shape;
 
 public interface Shape {
     void getShapeName(String shapeName); // zwraca nazwę figury ("square", "triangle" oraz "circle"),
-    void getField(double field); // zwraca obliczone pole powierzchni tej figury.
+    void getField(int field); // zwraca obliczone pole powierzchni tej figury.
 
 
 }
